Extract EC2 instance logic to machine.go
